Move WithPropertyAttribute to generator
